+++ rio_tiler/reader.py
@@ -10,7 +10,7 @@ import numpy
 from affine import Affine
 from rasterio import windows
 from rasterio.crs import CRS
-from rasterio.enums import ColorInterp, Resampling
+from rasterio.enums import ColorInterp, MaskFlags, Resampling
 from rasterio.io import DatasetReader, DatasetWriter
 from rasterio.vrt import WarpedVRT
 from rasterio.warp import transform as transform_coords
@@ -19,14 +19,7 @@ from rasterio.warp import transform_bounds
 from rio_tiler.constants import WGS84_CRS
 from rio_tiler.errors import InvalidBufferSize, PointOutsideBounds, TileOutsideBounds
 from rio_tiler.models import ImageData, PointData
-from rio_tiler.types import (
-    BBox,
-    DataMaskType,
-    Indexes,
-    NoData,
-    RIOResampling,
-    WarpResampling,
-)
+from rio_tiler.types import BBox, Indexes, NoData, RIOResampling, WarpResampling
 from rio_tiler.utils import _requested_tile_aligned_with_internal_tile as is_aligned
 from rio_tiler.utils import get_vrt_transform, has_alpha_band, non_alpha_indexes
 
@@ -40,7 +33,7 @@ class Options(TypedDict, total=False):
     resampling_method: Optional[RIOResampling]
     reproject_method: Optional[WarpResampling]
     unscale: Optional[bool]
-    post_process: Optional[Callable[[numpy.ndarray, numpy.ndarray], DataMaskType]]
+    post_process: Optional[Callable[[numpy.ma.MaskedArray], numpy.ma.MaskedArray]]
 
 
 def _get_width_height(max_size, dataset_height, dataset_width) -> Tuple[int, int]:
@@ -99,7 +92,7 @@ def read(
     reproject_method: WarpResampling = "nearest",
     unscale: bool = False,
     post_process: Optional[
-        Callable[[numpy.ndarray, numpy.ndarray], DataMaskType]
+        Callable[[numpy.ma.MaskedArray], numpy.ma.MaskedArray]
     ] = None,
 ) -> ImageData:
     """Low level read function.
@@ -197,7 +190,9 @@ def read(
                 resampling=io_resampling,
                 boundless=boundless,
             )
-            data, mask = data[0:-1], data[-1].astype("uint8")
+            mask = ~data[-1].astype("bool")
+            data = numpy.ma.MaskedArray(data[0:-1])
+            data.mask = mask
 
         else:
             data = dataset.read(
@@ -206,13 +201,12 @@ def read(
                 out_shape=(len(indexes), height, width) if height and width else None,
                 resampling=io_resampling,
                 boundless=boundless,
+                masked=True,
             )
-            mask = dataset.dataset_mask(
-                window=window,
-                out_shape=(height, width) if height and width else None,
-                resampling=io_resampling,
-                boundless=boundless,
-            )
+
+            # if data has Nodata then we simply make sure the mask == the nodata
+            if dataset.nodata is not None:
+                data.mask |= data == dataset.nodata
 
         stats = []
         for ix in indexes:
@@ -227,8 +221,10 @@ def read(
         # We only add dataset statistics if we have them for all the indexes
         dataset_statistics = stats if len(stats) == len(indexes) else None
 
+        # TODO: DEPRECATED, masked array are already using bool
         if force_binary_mask:
-            mask = numpy.where(mask != 0, numpy.uint8(255), numpy.uint8(0))
+            pass
+            # mask = numpy.where(mask != 0, numpy.uint8(255), numpy.uint8(0))
 
         if unscale:
             data = data.astype("float32", casting="unsafe")
@@ -236,22 +232,19 @@ def read(
             numpy.add(data, dataset.offsets[0], out=data, casting="unsafe")
 
         if post_process:
-            data, mask = post_process(data, mask)
+            data = post_process(data)
 
         out_bounds = (
             windows.bounds(window, dataset.transform) if window else dataset.bounds
         )
 
-        img = ImageData(
+        return ImageData(
             data,
-            mask,
             bounds=out_bounds,
             crs=dataset.crs,
             band_names=[f"b{idx}" for idx in indexes],
             dataset_statistics=dataset_statistics,
         )
-
-    return img
 
 
 # flake8: noqa: C901
@@ -274,7 +267,7 @@ def part(
     reproject_method: WarpResampling = "nearest",
     unscale: bool = False,
     post_process: Optional[
-        Callable[[numpy.ndarray, numpy.ndarray], DataMaskType]
+        Callable[[numpy.ma.MaskedArray], numpy.ma.MaskedArray]
     ] = None,
 ) -> ImageData:
     """Read part of a dataset.
@@ -425,9 +418,9 @@ def part(
             unscale=unscale,
             post_process=post_process,
         )
+
         return ImageData(
-            data=img.data[:, padding:-padding, padding:-padding],
-            mask=img.mask[padding:-padding, padding:-padding],
+            img.array[:, padding:-padding, padding:-padding],
             bounds=bounds,
             crs=img.crs,
             band_names=img.band_names,
@@ -460,7 +453,7 @@ def point(
     reproject_method: WarpResampling = "nearest",
     unscale: bool = False,
     post_process: Optional[
-        Callable[[numpy.ndarray, numpy.ndarray], DataMaskType]
+        Callable[[numpy.ma.MaskedArray], numpy.ma.MaskedArray]
     ] = None,
 ) -> PointData:
     """Read a pixel value for a point.
@@ -531,8 +524,7 @@ def point(
         )
 
     return PointData(
-        img.data[:, 0, 0],
-        numpy.array([img.mask[0, 0]]),
+        img.array[:, 0, 0],
         coordinates=coordinates,
         crs=coord_crs,
         band_names=img.band_names,